requirements.txt
# EcomTrendAI Dependencies

# HTTP/スクレイピング
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Amazon PA-API
python-amazon-paapi>=1.0.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# 設定管理
python-dotenv>=1.0.0

# 将来のML機能
# prophet>=1.1.0
# scikit-learn>=1.3.0
# torch>=2.0.0

# 開発・テスト
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# ログ
loguru>=0.7.0
